upgrade to ocotber v3

+++ modules/cms/classes/Asset.php
@@ -4,14 +4,14 @@ use File;
 use Lang;
 use Config;
 use Request;
-use ApplicationException;
-use ValidationException;
 use Cms\Helpers\File as FileHelper;
 use October\Rain\Extension\Extendable;
-use October\Rain\Filesystem\PathResolver;
+use ApplicationException;
+use ValidationException;
+use DirectoryIterator;
 
 /**
- * The CMS theme asset file class.
+ * Asset for CMS asset files
  *
  * @package october\cms
  * @author Alexey Bobkov, Samuel Georges
@@ -24,12 +24,12 @@ class Asset extends Extendable
     protected $theme;
 
     /**
-     * @var string The container name inside the theme.
+     * @var string dirName for the container name inside the theme
      */
     protected $dirName = 'assets';
 
     /**
-     * @var string Specifies the file name corresponding the CMS object.
+     * @var string fileName specifies the file name corresponding the CMS object
      */
     public $fileName;
 
@@ -49,7 +49,7 @@ class Asset extends Extendable
     public $content;
 
     /**
-     * @var array The attributes that are mass assignable.
+     * @var array fillable attributes that are mass assignable
      */
     protected $fillable = [
         'fileName',
@@ -57,18 +57,17 @@ class Asset extends Extendable
     ];
 
     /**
-     * @var array Allowable file extensions.
+     * @var array allowedExtensions for template files
      */
     protected $allowedExtensions = [];
 
     /**
-     * @var bool Indicates if the model exists.
+     * @var bool exists indicates if the model exists.
      */
     public $exists = false;
 
     /**
-     * Creates an instance of the object and associates it with a CMS theme.
-     * @param \Cms\Classes\Theme $theme Specifies the theme the object belongs to.
+     * __construct creates an instance of the object and associates it with a CMS theme
      */
     public function __construct(Theme $theme)
     {
@@ -80,7 +79,7 @@ class Asset extends Extendable
     }
 
     /**
-     * Loads the object from a file.
+     * load the object from a file
      * This method is used in the CMS back-end. It doesn't use any caching.
      * @param \Cms\Classes\Theme $theme Specifies the theme the object belongs to.
      * @param string $fileName Specifies the file name, with the extension.
@@ -93,11 +92,136 @@ class Asset extends Extendable
     }
 
     /**
-     * Prepares the theme datasource for the model.
-     * @param \Cms\Classes\Theme|string $theme Specifies a parent theme.
-     * @return $this
+     * listInTheme
      */
-    public static function inTheme($theme)
+    public static function listInTheme($theme, array $options = [])
+    {
+        return static::inTheme($theme)->get($options);
+    }
+
+    /**
+     * get all assets in a theme and uses simple objects
+     *
+     * Available options:
+     * - recursive: search subfolders and place in 'assets' key
+     * - flatten: produce a flat array instead of a recursive array
+     * - filterPath: only include within an inner path
+     * - filterFiles: only include files
+     * - filterFolders: only include folders
+     * - filterEditable: only show editable assets
+     */
+    public function get(array $options = []): array
+    {
+        extract(array_merge([
+            'recursive' => true,
+            'flatten' => false,
+            'filterPath' => '',
+            'filterFiles' => false,
+            'filterFolders' => false,
+            'filterEditable' => false,
+        ], $options));
+
+        $assets = [];
+
+        $pathSuffix = $filterPath ? '/'.$filterPath : '';
+        $path = $this->theme->getPath().'/'.$this->dirName.$pathSuffix;
+        $files = $this->getInternal($path, $this->theme);
+
+        // Splice in assets of parent theme
+        if ($parentTheme = $this->theme->getParentTheme()) {
+            $parentPath = $parentTheme->getPath().'/'.$this->dirName.$pathSuffix;
+            $files = array_merge($files, $this->getInternal($parentPath, $parentTheme));
+        }
+
+        foreach ($files as $asset) {
+            if ($recursive && $asset['isFolder'] && $asset['filename']) {
+                $newFilter = $pathSuffix ? $pathSuffix.'/'.$asset['filename'] : $asset['filename'];
+
+                if ($flatten) {
+                    $assets = array_merge($assets, $this->get(['filterPath' => $newFilter] + $options));
+                }
+                else {
+                    $asset['assets'] = $this->get(['filterPath' => $newFilter] + $options);
+                }
+            }
+
+            if ($filterFolders && !$asset['isFolder']) {
+                continue;
+            }
+
+            if ($filterEditable && !$asset['isEditable']) {
+                continue;
+            }
+
+            if ($filterFiles && $asset['isFolder']) {
+                continue;
+            }
+
+            $assets[] = $asset;
+        }
+
+        return collect($assets)->keyBy('path')->all();
+    }
+
+    /**
+     * getInternal helps the get method
+     */
+    protected function getInternal(string $path, Theme $theme): array
+    {
+        if (!file_exists($path)) {
+            return [];
+        }
+
+        $result = [];
+        $iterator = new DirectoryIterator($path);
+        $editableAssetTypes = Asset::getEditableExtensions();
+
+        foreach ($iterator as $fileInfo) {
+            $fileName = $fileInfo->getFileName();
+            if (substr($fileName, 0, 1) === '.') {
+                continue;
+            }
+
+            if (!$fileInfo->isDir() && !$fileInfo->isFile()) {
+                continue;
+            }
+
+            $fileName = $fileInfo->getFileName();
+            $isFolder = $fileInfo->isDir();
+            $filePath = $this->getRelativePath($fileInfo->getPathname(), $theme);
+            $isEditable = in_array(strtolower($fileInfo->getExtension()), $editableAssetTypes);
+
+            $asset = [
+                'filename' => $fileName,
+                'isFolder' => $isFolder ? 1 : 0,
+                'isEditable' => $isEditable,
+                'path' => ltrim(File::normalizePath($filePath), '/')
+            ];
+
+            $result[] = $asset;
+        }
+
+        return $result;
+    }
+
+    /**
+     * getRelativePath returns path relative to the theme asset directory
+     */
+    protected function getRelativePath(string $path, Theme $theme): string
+    {
+        $prefix = $theme->getPath().'/'.$this->dirName;
+
+        if (substr($path, 0, strlen($prefix)) === $prefix) {
+            $path = substr($path, strlen($prefix));
+        }
+
+        return $path;
+    }
+
+    /**
+     * inTheme prepares the theme datasource for the model.
+     */
+    public static function inTheme($theme): Asset
     {
         if (is_string($theme)) {
             $theme = Theme::load($theme);
@@ -107,17 +231,35 @@ class Asset extends Extendable
     }
 
     /**
-     * Find a single template by its file name.
-     *
-     * @param  string $fileName
-     * @return mixed|static
+     * find a single template by its file name.
      */
-    public function find($fileName)
+    public function find(string $fileName)
     {
         $filePath = $this->getFilePath($fileName);
 
+        $foundTheme = $this->theme;
+
         if (!File::isFile($filePath)) {
-            return null;
+            // Look at parent
+            if ($parentTheme = $this->theme->getParentTheme()) {
+                $foundTheme = $parentTheme;
+                $filePath = $parentTheme->getPath().'/'.$this->dirName.'/'.$fileName;
+
+                if (!File::isFile($filePath)) {
+                    return null;
+                }
+            }
+            else {
+                return null;
+            }
+        }
+
+        if (!FileHelper::validateInTheme($foundTheme, $filePath)) {
+            throw new ValidationException(['fileName' =>
+                Lang::get('cms::lang.cms_object.invalid_file', [
+                    'name' => $fileName
+                ])
+            ]);
         }
 
         if (($content = @File::get($filePath)) === false) {
@@ -129,6 +271,7 @@ class Asset extends Extendable
         $this->mtime = File::lastModified($filePath);
         $this->content = $content;
         $this->exists = true;
+
         return $this;
     }
 
@@ -151,7 +294,7 @@ class Asset extends Extendable
     }
 
     /**
-     * Saves the object to the disk.
+     * save the object to the disk
      */
     public function save()
     {
@@ -168,7 +311,7 @@ class Asset extends Extendable
 
         $dirPath = $this->theme->getPath().'/'.$this->dirName;
         if (!file_exists($dirPath) || !is_dir($dirPath)) {
-            if (!File::makeDirectory($dirPath, 0777, true, true)) {
+            if (!File::makeDirectory($dirPath, 0755, true, true)) {
                 throw new ApplicationException(Lang::get(
                     'cms::lang.cms_object.error_creating_directory',
                     ['name'=>$dirPath]
@@ -179,7 +322,7 @@ class Asset extends Extendable
         if (($pos = strpos($this->fileName, '/')) !== false) {
             $dirPath = dirname($fullPath);
 
-            if (!is_dir($dirPath) && !File::makeDirectory($dirPath, 0777, true, true)) {
+            if (!is_dir($dirPath) && !File::makeDirectory($dirPath, 0755, true, true)) {
                 throw new ApplicationException(Lang::get(
                     'cms::lang.cms_object.error_creating_directory',
                     ['name'=>$dirPath]
@@ -210,12 +353,23 @@ class Asset extends Extendable
         $this->exists = true;
     }
 
+    /**
+     * delete the object from disk
+     */
     public function delete()
     {
         $fileName = Request::input('fileName');
         $fullPath = $this->getFilePath($fileName);
 
         $this->validateFileName($fileName);
+
+        if (!FileHelper::validateInTheme($this->theme, $fullPath)) {
+            throw new ValidationException(['fileName' =>
+                Lang::get('cms::lang.cms_object.invalid_file', [
+                    'name' => $fileName
+                ])
+            ]);
+        }
 
         if (File::exists($fullPath)) {
             if (!@File::delete($fullPath)) {
@@ -228,7 +382,7 @@ class Asset extends Extendable
     }
 
     /**
-     * Validate the supplied filename, extension and path.
+     * validateFileName supplied with extension and path.
      * @param string $fileName
      */
     protected function validateFileName($fileName = null)
@@ -267,7 +421,15 @@ class Asset extends Extendable
     }
 
     /**
-     * Returns the file name.
+     * validate object
+     */
+    public function validate()
+    {
+        $this->validateFileName();
+    }
+
+    /**
+     * getFileName
      * @return string
      */
     public function getFileName()
@@ -276,37 +438,27 @@ class Asset extends Extendable
     }
 
     /**
-     * Returns the absolute file path.
-     * @param string $fileName Specifies the file name to return the path to.
-     * @return string
+     * getFilePath returns the absolute file path of an asset
      */
-    public function getFilePath($fileName = null)
+    public function getFilePath(string $fileName = null): string
     {
         if ($fileName === null) {
             $fileName = $this->fileName;
         }
 
-        $directory = $this->theme->getPath() . '/' . $this->dirName . '/';
-        $filePath = $directory . $fileName;
-
-        // Limit paths to those under the theme's assets directory
-        if (!PathResolver::within($filePath, $directory)) {
-            return false;
-        }
-
-        return PathResolver::resolve($filePath);
+        return $this->theme->getPath().'/'.$this->dirName.'/'.$fileName;
     }
 
     /**
-     * Returns a list of editable asset extensions.
-     * The list can be overridden with the cms.editableAssetTypes configuration option.
+     * getEditableExtensions returns a list of editable asset extensions
+     * The list can be overridden with the cms.editable_asset_types configuration option.
      * @return array
      */
     public static function getEditableExtensions()
     {
         $defaultTypes =  ['css', 'js', 'less', 'sass', 'scss'];
 
-        $configTypes = Config::get('cms.editableAssetTypes');
+        $configTypes = Config::get('cms.editable_asset_types');
         if (!$configTypes) {
             return $defaultTypes;
         }
